+++ libs/next/src/lib/components/chips/chips.component.ts
@@ -1,7 +1,9 @@
 import {
+  AfterViewInit,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
+  ElementRef,
   EventEmitter,
   forwardRef,
   HostBinding,
@@ -11,9 +13,11 @@ import {
   Output,
 } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { BehaviorSubject, Subscription } from 'rxjs';
 import { zuiIsTextOverflow$ } from '../../util/dom/is-textoverflow';
 import { ZuiOverlayOutsidePlacement } from '../../modules';
+import { BehaviorSubject, Observable, of, Subscription, timer } from 'rxjs';
+import { map, switchMap, takeUntil } from 'rxjs/operators';
+import { ZuiDestroyService } from '@digital-plant/zyfra-helpers';
 
 @Component({
   selector: 'zui-chips',
@@ -26,14 +30,16 @@ import { ZuiOverlayOutsidePlacement } from '../../modules';
       useExisting: forwardRef(() => ZuiChipsComponent),
       multi: true,
     },
+    ZuiDestroyService
   ],
 })
-export class ZuiChipsComponent implements ControlValueAccessor, OnInit, OnDestroy {
+export class ZuiChipsComponent implements ControlValueAccessor, OnInit, OnDestroy, AfterViewInit {
   @Input() @HostBinding('attr.data-size') public size: 's' | 'l' = 'l';
   @Input() set chips(data: string[]) {
     this.chipsList = data;
   }
   @Input() public deletable = true;
+  @Input() public singleLine = true;
   @Input() public hintCanShow = true;
   @Input() public hintDirection: ZuiOverlayOutsidePlacement = ZuiOverlayOutsidePlacement.RIGHT;
   @Input() set disabled(isDisabled: boolean) {
@@ -46,14 +52,35 @@ export class ZuiChipsComponent implements ControlValueAccessor, OnInit, OnDestro
 
   @HostBinding('attr.testId')
   readonly testId = 'zui_chips';
-
-
-  readonly zuiIsTextOverflow$ = zuiIsTextOverflow$;
   public accessorIsDisabled = false;
+  public readonly overflowedChipsList$ = new BehaviorSubject<Set<number>>(new Set());
+
   public chipsList$: BehaviorSubject<string[]> = new BehaviorSubject<string[]>([]);
   private subscription: Subscription = new Subscription();
+  readonly ready: (el: ElementRef) => boolean = (el: ElementRef) => {
+    const {x, y} = el.nativeElement.getBoundingClientRect();
+    return Math.max(x, y) > 0
+  }
 
-  constructor(private readonly cdRef: ChangeDetectorRef) {}
+  constructor(private readonly cdRef: ChangeDetectorRef, private readonly destroy$: ZuiDestroyService) {}
+
+  readonly zuiIsTextOverflow$ = (elem: HTMLElement, hintCanShow: boolean, forceShowHint: boolean): Observable<boolean> => {
+    return of(forceShowHint).pipe(
+      switchMap(
+        val => {
+          if (val) {
+            return of(true);
+          }
+
+          if (!hintCanShow) {
+            return of(false);
+          }
+
+          return zuiIsTextOverflow$(elem);
+        }
+      )
+    )
+  };
 
   // eslint-disable-next-line @typescript-eslint/no-empty-function
   public onChange: (value: unknown) => void = () => {};
@@ -77,6 +104,9 @@ export class ZuiChipsComponent implements ControlValueAccessor, OnInit, OnDestro
 
   public removeChips(event: MouseEvent, idx: number): void {
     if (this.accessorIsDisabled) return;
+    this.overflowedChipsList$.value.delete(idx);
+    this.overflowedChipsList$.next(this.overflowedChipsList$.value);
+
     event.stopPropagation();
     this.removeChipEvent.emit(this.chipsList[idx]);
     this.chipsList = this.chipsList.filter((item, i) => i !== idx);
@@ -88,7 +118,9 @@ export class ZuiChipsComponent implements ControlValueAccessor, OnInit, OnDestro
   }
 
   public ngOnInit(): void {
-    this.subscription = this.chipsList$.subscribe(chips => {
+    this.subscription = this.chipsList$.pipe(
+      takeUntil(this.destroy$)
+    ).subscribe(chips => {
       this.onChange(chips);
     });
   }
@@ -109,8 +141,66 @@ export class ZuiChipsComponent implements ControlValueAccessor, OnInit, OnDestro
   public registerOnChange(fn: () => void): void {
     this.onChange = fn;
   }
+
   public registerOnTouched(fn: () => void): void {
     this.onTouched = fn;
+  }
+
+
+  public isChipsContent$ (
+    observable: Observable<ElementRef>,
+    parent: HTMLElement,
+    singleLine: boolean,
+    chips: string,
+    idx: number,
+    allChipsCount: number
+  ): Observable<string> {
+    return this.chipsList$.pipe(
+      switchMap(
+        () => observable
+      ),
+      map((current: ElementRef) => {
+        if (idx === 0) this.overflowedChipsList$.value.clear();
+        if (!singleLine || this.chipsList.length === 1 ) return false;
+
+        const maxPadding = 2;
+        const needWidthPlaceForShowDots = 35;
+        const offsetY = Math.abs((parent.offsetTop - current.nativeElement.offsetTop)) > maxPadding;
+
+        const parentX = parent.offsetLeft + parent.offsetWidth;
+        const currentX = current.nativeElement.offsetLeft + current.nativeElement.offsetWidth;
+        const result = offsetY ||
+          (parentX - currentX) < needWidthPlaceForShowDots;
+
+        if (result)
+          this.overflowedChipsList$.value.add(idx);
+        else
+          this.overflowedChipsList$.value.delete(idx);
+
+        this.overflowedChipsList$.next(
+          this.overflowedChipsList$.value
+        );
+
+        return result;
+      }),
+      map(i => i ? 'hidden' : 'visible')
+    )
+  }
+
+  public getOverflowedChipsListHint(
+  ): string {
+    const list = [...this.overflowedChipsList$.value.values()];
+    return [...list].map(i => {
+      return this.chipsList[i];
+    }).join(', ')
+  }
+
+  ngAfterViewInit(): void {
+    timer(0).subscribe(
+      () => {
+        this.chipsList = [...this.chipsList];
+      }
+    )
   }
 }