+++ CHANGELOG.md
@@ -4,6 +4,34 @@
 
 ---
 
+## [v0.13.8] - 18-10-2022
+
+### @digital-plant/zyfra-components [v0.13.8]
+
+**Changes**
+
+- fix/(components/package): fix peerDependencies [MR](https://gitdp.zyfra.com/digital-plant/ziiot/ui-platform/frontend/zui-sdk/-/merge_requests/353)
+
+> **Special thanks**
+>
+> ✏️ **MR:** Zurab Magomadov
+
+
+---
+
+## [1.0.0-beta.10](https://gitdp.zyfra.com/digital-plant/ziiot/ui-platform/frontend/zui-sdk) (18-10-2022)
+
+### Features
+- feat(next/icon): add poi svg icons [MR](https://gitdp.zyfra.com/digital-plant/ziiot/ui-platform/frontend/zui-sdk/-/merge_requests/333)
+- feat(next/multi-select): add single line chips multi-select [MR](https://gitdp.zyfra.com/digital-plant/ziiot/ui-platform/frontend/zui-sdk/-/merge_requests/353)
+- feat(next/chips): add single line for chips [MR](https://gitdp.zyfra.com/digital-plant/ziiot/ui-platform/frontend/zui-sdk/-/merge_requests/353)
+
+### BUG FIXES
+- feat(next/chips): fix select border [MR](https://gitdp.zyfra.com/digital-plant/ziiot/ui-platform/frontend/zui-sdk/-/merge_requests/353)
+
+
+---
+
 ## [v0.13.7] - 14-10-2022
 
 ### @digital-plant/zyfra-components [v0.13.7]